Asynchronous Examples refined
Created a new example where we show how a planner can orchestrate the actions of other planners.

Refined the code for the test_async.py to be cleaner, more well documented

Reinforced the logic of register_agent where it distinguishes between basic agents or planner agents when giving discriptions.

Removed the debug statement and added logs as a native part of the implementation.